spltmProject/requirements.txt
# Django & REST Framework
Django==4.2.11
djangorestframework==3.14.0

# Database Drivers
mysqlclient==2.2.4
mariadb==1.1.14
mssql-django==1.6
pyodbc==5.3.0

# Utilities
asgiref==3.11.0
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3

# Authentication & Security
PyJWT==2.10.1

# Image Processing
pillow==10.4.0

# Type Hints
typing_extensions==4.15.0

# Package Management
packaging==25.0

setuptools==80.9.0
wheel==0.45.1
